pyproject.toml
[project]
name = "onvif-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onvif-zeep>=0.2.12",
]
